First Deliverable over

+++ src/com/lteu/deliverable/Location.java
@@ -3,28 +3,28 @@ package com.lteu.deliverable;
 
 
 public class Location {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Location(int x, int y) {
+    public Location(double x, double y) {
         super();
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
